fixed component structure
