+++ gotexplorer.client/src/main.tsx
@@ -10,20 +10,21 @@ import ForgetPasswordPage from './pages/auth/ForgetPassword.tsx';
 import SetNewPasswordPage from './pages/auth/SetNewPassword.tsx';
 import ProfilePage from './pages/auth/ProfilePage.tsx';
 import StartGamePage from './pages/games/StartGamePage.tsx';
-import GoToGame from './pages/games/GoToGame.tsx';
+import Layout from './pages/additional_components/Layout.tsx';
 
 export default function App() {
     return (
         <BrowserRouter>
             <Routes>
                 <Route index element={<MainPage />} />
+                <Route path="/" element=<Layout />>
+                    <Route path="profile" element={<ProfilePage />} />
+                    <Route path="startgame" element={<StartGamePage />} />
+                </Route>
                 <Route path="signup" element={<SignUpPage />} />
                 <Route path="login" element={<LogInPage />} />
                 <Route path="forgetpass" element={<ForgetPasswordPage />} />
                 <Route path="setnewpass" element={<SetNewPasswordPage />} />
-                <Route path="profile" element={<ProfilePage />} />
-                <Route path="startgame" element={<StartGamePage />} />
-                <Route path="gotogame" element={<GoToGame />} />
             </Routes>
         </BrowserRouter>
     );
