+++ gotexplorer.client/src/pages/additional_components/Layout.tsx
@@ -1,0 +1,14 @@
+import { Outlet } from "react-router-dom";
+import Navigation from "./Navigation";
+import Footer from "./Footer";
+
+const Layout = () => {
+    return (<>
+        <div style={{ backgroundColor: "black" }}>
+            <Navigation />
+        </div>
+        <Outlet />
+        <Footer />
+    </>);
+}
+export default Layout;
