--- gotexplorer.client/src/pages/games/GoToGame.tsx
@@ -1,14 +1,0 @@
-import Footer from "../additional_components/Footer";
-import Navigation from "../additional_components/Navigation";
-
-const GoToGame = () => {
-    return (<>
-        <div style={{ backgroundColor: "black" }}>
-            <Navigation />
-        </div>
-        <h2>Go To Game</h2>
-        <Footer />
-    </>
-    );
-}
-export default GoToGame;
